feat: several improvements so that the model now trains correctly
In order for the model to learn the effects of it's action, it is necessary that a certain amount of time passes between the model's action and getting the resulting next state. Method tf.nextFrame() was not suitable for this task and so this method is replaced by a delay loop based on the game's own timer. In addition, optimization step was moved into this interval between state and next state to best use the available time here.

Game state was fetched using several separate methods. This led to inconsistencies due to time passing between calls to those methods. Instead, there now is a single method to get the complete game state at once consistently.w